refactor: move storage owner boundary aggregation
